fix: unbreak macOS build and correct the reader examples
macOS CI has been red since the runner image moved to Xcode 26.6 —
including on docs-only commits to develop, so this predates this branch.
clang now rejects `UINT64_MAX * sample_rate` and `... / UINT64_MAX` under
-Werror, because UINT64_MAX has no exact double representation
(-Wimplicit-const-int-float-conversion). Made both widenings explicit in
mrcProfiler; the arithmetic is unchanged, and both SHARDS paths were
re-run to confirm identical output.

doc/advanced_lib.md's reader examples were wrong in three ways, which
repointing them at the sample trace made visible:

- obj_id_field was 6, but cloudPhysicsIO.csv is version,time,op,size,lbn
  so the id is field 5. Copying it verbatim gave every row the same
  default id and meaningless results.
- has_header was FALSE for a file that has a header, and obj_id_is_num
  was unset on a numeric id column.
- the binary example used .binary_fmt, which is not a field
  (.binary_fmt_str), with "<3I2H2Q" — the format parser has no repeat
  counts, so it errored with "unknown format '3'". Expanded to
  "<IIIHHQQ", matching the CLI docs.

Both snippets were compiled and run against the sample traces: the csv
and binary readers now agree with the vscsi reader exactly — 113872
requests, miss ratio 0.6297, same first object.

While in SLRU_parse_params, validated the parameters it never checked.
`-e n-seg=0` divided by zero in init (SIGFPE, present on develop), and
seg-size wrote past seg_size_array[SLRU_MAX_N_SEG] once given more than
16 segments. Both now report what is wrong.

+++ libCacheSim/mrcProfiler/mrcProfiler.cpp
@@ -95,7 +95,10 @@ void mrcProfiler::MRCProfilerSHARDS::fixed_sample_rate_run() {
   double sample_rate = params_.shards_params.sample_rate;
   std::vector<double> local_hit_cnt_vec(mrc_size_vec.size(), 0);
   std::vector<double> local_hit_size_vec(mrc_size_vec.size(), 0);
-  uint64_t sample_max = UINT64_MAX * sample_rate;
+  /* UINT64_MAX has no exact double representation, so make the widening
+   * explicit; clang errors on the implicit form under -Werror */
+  uint64_t sample_max =
+      static_cast<uint64_t>(static_cast<double>(UINT64_MAX) * sample_rate);
   if (sample_rate == 1) {
     INFO("sample_rate is 1, no need to sample\n");
     sample_max = UINT64_MAX;
@@ -212,8 +215,9 @@ void mrcProfiler::MRCProfilerSHARDS::fixed_sample_size_run() {
       if (!min_value_map.full()) {
         sample_rate = 1.0;  // still 100% sample rate
       } else {
-        sample_rate = min_value_map.get_max_value() * 1.0 /
-                      UINT64_MAX;  // adjust the sample rate
+        sample_rate =
+            min_value_map.get_max_value() * 1.0 /
+            static_cast<double>(UINT64_MAX);  // adjust the sample rate
       }
 
       sampled_cnt += 1.0 / sample_rate;
